# accounts/src/main/java/avlyakulov/timur/accounts/dto/ResponseDto.java
package avlyakulov.timur.accounts.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class ResponseDto {

    private String statusCode;

    private String statusMsg;
}